fix(ci): sync mobile locales and resolve iOS SwiftLint errors
- Regenerate Localizable.xcstrings from sync-mobile-locales.py
- Split LMSAPI/LMSFeatureModels additions into mobile wave files to stay under file_length limits
- Fix identifier_name, cyclomatic_complexity, and type_body_length violations
- Extract ProfileView and DashboardView sections to reduce struct body size
