+++ clients/ios/Lextures/Core/LMS/LMSAPIFeatures.swift
@@ -15,16 +15,6 @@ extension LMSAPI {
         return try decode(MeProfile.self, from: data)
     }
 
-    static func fetchMyPermissions(accessToken: String) async throws -> [String] {
-        let (data, _) = try await client.request(
-            path: "/api/v1/me/permissions",
-            authorized: true,
-            accessToken: accessToken
-        )
-        let response = try decode(MyPermissionsResponse.self, from: data)
-        return response.permissionStrings
-    }
-
     // MARK: - Account settings (editable profile)
 
     static func fetchAccountProfile(accessToken: String) async throws -> AccountProfile {
@@ -59,87 +49,6 @@ extension LMSAPI {
             accessToken: accessToken
         )
         return try decode(MyAccommodationsResponse.self, from: data).accommodations
-    }
-
-    // MARK: - Profile depth (M1.5)
-
-    static func fetchMyProfileFields(accessToken: String) async throws -> ProfileFieldsResponse {
-        let (data, _) = try await client.request(
-            path: "/api/v1/me/profile-fields",
-            authorized: true,
-            accessToken: accessToken
-        )
-        return try decode(ProfileFieldsResponse.self, from: data)
-    }
-
-    static func updateMyProfileFields(
-        _ patch: ProfileFieldsPatch,
-        accessToken: String
-    ) async throws -> [String: JSONValue] {
-        let (data, _) = try await client.request(
-            path: "/api/v1/me/profile-fields",
-            method: "PATCH",
-            body: patch,
-            authorized: true,
-            accessToken: accessToken
-        )
-        let response = try decode(ProfileFieldsValuesResponse.self, from: data)
-        return response.values
-    }
-
-    static func fetchMyDemographics(accessToken: String) async throws -> StudentDemographics {
-        let (data, _) = try await client.request(
-            path: "/api/v1/me/demographics",
-            authorized: true,
-            accessToken: accessToken
-        )
-        return try decode(StudentDemographics.self, from: data)
-    }
-
-    static func updateMyDemographics(
-        _ patch: StudentDemographicsPatch,
-        accessToken: String
-    ) async throws -> StudentDemographics {
-        let (data, _) = try await client.request(
-            path: "/api/v1/me/demographics",
-            method: "PATCH",
-            body: patch,
-            authorized: true,
-            accessToken: accessToken
-        )
-        return try decode(StudentDemographics.self, from: data)
-    }
-
-    static func fetchPendingConsentStudies(accessToken: String) async throws -> [ConsentStudy] {
-        let (data, _) = try await client.request(
-            path: "/api/v1/me/consent-studies",
-            authorized: true,
-            accessToken: accessToken
-        )
-        return try decode(ConsentStudiesResponse.self, from: data).studies
-    }
-
-    static func fetchConsentHistory(accessToken: String) async throws -> [ConsentHistoryEntry] {
-        let (data, _) = try await client.request(
-            path: "/api/v1/me/consent-studies/history",
-            authorized: true,
-            accessToken: accessToken
-        )
-        return try decode(ConsentHistoryResponse.self, from: data).history
-    }
-
-    static func respondToConsentStudy(
-        studyId: String,
-        decision: ConsentDecision,
-        accessToken: String
-    ) async throws {
-        _ = try await client.request(
-            path: "/api/v1/me/consent-studies/\(encodePath(studyId))/respond",
-            method: "POST",
-            body: ConsentRespondBody(decision: decision),
-            authorized: true,
-            accessToken: accessToken
-        )
     }
 
     // MARK: - Notifications
@@ -613,78 +522,6 @@ extension LMSAPI {
         )
     }
 
-    static func createAttendanceSession(
-        courseCode: String,
-        body: CreateAttendanceSessionBody,
-        accessToken: String
-    ) async throws -> AttendanceSession {
-        let (data, response) = try await client.request(
-            path: "/api/v1/courses/\(encodePath(courseCode))/attendance/sessions",
-            method: "POST",
-            body: body,
-            authorized: true,
-            accessToken: accessToken
-        )
-        guard (200 ... 299).contains(response.statusCode) else {
-            throw APIError.httpStatus(response.statusCode, message: parseAPIErrorMessage(from: data))
-        }
-        return try decode(AttendanceSession.self, from: data)
-    }
-
-    static func saveAttendanceRecords(
-        courseCode: String,
-        sessionId: String,
-        records: [AttendanceRecordUpsert],
-        accessToken: String
-    ) async throws -> SaveAttendanceRecordsResponse {
-        let (data, response) = try await client.request(
-            path: "/api/v1/courses/\(encodePath(courseCode))/attendance/sessions/\(encodePath(sessionId))/records",
-            method: "PUT",
-            body: SaveAttendanceRecordsBody(records: records),
-            authorized: true,
-            accessToken: accessToken
-        )
-        guard (200 ... 299).contains(response.statusCode) else {
-            throw APIError.httpStatus(response.statusCode, message: parseAPIErrorMessage(from: data))
-        }
-        return try decode(SaveAttendanceRecordsResponse.self, from: data)
-    }
-
-    struct CloseAttendanceSessionBody: Encodable {
-        var finalizeMissingAsAbsent: Bool = true
-    }
-
-    static func closeAttendanceSession(
-        courseCode: String,
-        sessionId: String,
-        accessToken: String
-    ) async throws -> AttendanceSession {
-        let (data, response) = try await client.request(
-            path: "/api/v1/courses/\(encodePath(courseCode))/attendance/sessions/\(encodePath(sessionId))/close",
-            method: "POST",
-            body: CloseAttendanceSessionBody(),
-            authorized: true,
-            accessToken: accessToken
-        )
-        guard (200 ... 299).contains(response.statusCode) else {
-            throw APIError.httpStatus(response.statusCode, message: parseAPIErrorMessage(from: data))
-        }
-        return try decode(AttendanceSession.self, from: data)
-    }
-
-    static func fetchCourseSections(courseCode: String, accessToken: String) async throws -> [CourseSection] {
-        let (data, response) = try await client.request(
-            path: "/api/v1/courses/\(encodePath(courseCode))/sections",
-            authorized: true,
-            accessToken: accessToken
-        )
-        if response.statusCode == 404 { return [] }
-        guard (200 ... 299).contains(response.statusCode) else {
-            throw APIError.httpStatus(response.statusCode, message: parseAPIErrorMessage(from: data))
-        }
-        return try decode(CourseSectionsResponse.self, from: data).sections
-    }
-
     // MARK: - Onboarding (plan 15.11 / M1.3)
 
     /// Returns nil when the onboarding feature flag is off (HTTP 404).
@@ -1042,98 +879,4 @@ extension LMSAPI {
         return join?.isEmpty == false ? join : nil
     }
 
-    // MARK: - Universal search (M0.6)
-
-    static func fetchSearchIndex(accessToken: String) async throws -> SearchIndexResponse {
-        let (data, _) = try await client.request(
-            path: "/api/v1/search",
-            authorized: true,
-            accessToken: accessToken
-        )
-        return try decode(SearchIndexResponse.self, from: data)
-    }
-
-    static func fetchSearchQuery(
-        query: String,
-        scope: String? = nil,
-        accessToken: String
-    ) async throws -> SearchQueryResponse {
-        var path = "/api/v1/search/query?q=\(encodePath(query))"
-        if let scope, !scope.isEmpty {
-            path += "&scope=\(encodePath(scope))"
-        }
-        let (data, _) = try await client.request(
-            path: path,
-            authorized: true,
-            accessToken: accessToken
-        )
-        return try decode(SearchQueryResponse.self, from: data)
-    }
-
-    // MARK: - Library & OER (M3.6)
-
-    static func searchLibraryCatalog(query: String, accessToken: String) async throws -> [LibraryCatalogResult] {
-        let q = query.trimmingCharacters(in: .whitespacesAndNewlines)
-        let (data, _) = try await client.request(
-            path: "/api/v1/library/search?q=\(encodePath(q))",
-            authorized: true,
-            accessToken: accessToken
-        )
-        return try decode(LibrarySearchResponse.self, from: data).results
-    }
-
-    static func fetchOERProviders(accessToken: String) async throws -> [String] {
-        let (data, _) = try await client.request(
-            path: "/api/v1/oer/providers",
-            authorized: true,
-            accessToken: accessToken
-        )
-        let rows = try decode([OERProviderRow].self, from: data)
-        return rows.map(\.provider)
-    }
-
-    static func searchOER(
-        provider: String,
-        query: String,
-        accessToken: String
-    ) async throws -> OERSearchResponse {
-        var path = "/api/v1/oer/search?provider=\(encodePath(provider))"
-        let q = query.trimmingCharacters(in: .whitespacesAndNewlines)
-        if !q.isEmpty {
-            path += "&q=\(encodePath(q))"
-        }
-        let (data, _) = try await client.request(
-            path: path,
-            authorized: true,
-            accessToken: accessToken
-        )
-        return try decode(OERSearchResponse.self, from: data)
-    }
-
-    static func fetchModuleLibraryResource(
-        courseCode: String,
-        itemId: String,
-        accessToken: String
-    ) async throws -> LibraryResourcePayload? {
-        let (data, response) = try await client.request(
-            path: "/api/v1/courses/\(encodePath(courseCode))/library-resources/\(encodePath(itemId))",
-            authorized: true,
-            accessToken: accessToken
-        )
-        if response.statusCode == 404 { return nil }
-        return try decode(LibraryResourcePayload.self, from: data)
-    }
-
-    static func recordLibraryResourceAccess(
-        courseCode: String,
-        itemId: String,
-        accessToken: String
-    ) async throws {
-        _ = try await client.request(
-            path: "/api/v1/courses/\(encodePath(courseCode))/library-resources/\(encodePath(itemId))/access",
-            method: "POST",
-            authorized: true,
-            accessToken: accessToken
-        )
-    }
 }
