+++ clients/ios/Lextures/Features/Profile/ProfileView.swift
@@ -23,11 +23,11 @@ struct ProfileView: View {
                     VStack(alignment: .leading, spacing: 16) {
                         identityHero
                         if shell.iaRedesignEnabled {
-                            iaContextCard
-                            moreHubCard
+                            ProfileIaContextCard()
+                            ProfileMoreHubCard()
                         }
                         if offline.pendingCount > 0 {
-                            offlineSyncCard
+                            ProfileOfflineSyncCard()
                         }
                         ProfilePersonalCard()
                         ProfileDepthCards()
@@ -37,7 +37,7 @@ struct ProfileView: View {
                         accessibilityCard
                         ProfileSecurityCard()
                         accountCard
-                        notificationsCard
+                        ProfileNotificationsCard()
                         ProfileLegalCard()
                         aboutCard
                         signOutButton
@@ -72,10 +72,10 @@ struct ProfileView: View {
                 MoreHubView()
             }
             .navigationDestination(for: MoreDestination.self) { destination in
-                moreDestinationView(destination)
+                ProfileMoreDestinationScreen(destination: destination)
             }
             .navigationDestination(item: $navigatedMoreDestination) { destination in
-                moreDestinationView(destination)
+                ProfileMoreDestinationScreen(destination: destination)
             }
             .confirmationDialog(
                 L.text("mobile.profile.signOutConfirm"),
@@ -154,82 +154,6 @@ struct ProfileView: View {
 
     private var profileInitials: String {
         shell.accountProfile?.resolvedInitials ?? shell.profile?.initials ?? "··"
-    }
-
-    private var iaContextCard: some View {
-        Group {
-            if shell.roleSnapshot.availableContexts.count > 1 {
-                LMSCard {
-                    Text(L.text("mobile.ia.context.title"))
-                        .font(LexturesTheme.displayFont(17))
-                        .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
-                    Picker(L.text("mobile.ia.context.title"), selection: Binding(
-                        get: { shell.activeRoleContext },
-                        set: { shell.setRoleContext($0) }
-                    )) {
-                        ForEach(shell.roleSnapshot.availableContexts, id: \.self) { context in
-                            Text(context.label).tag(context)
-                        }
-                    }
-                    .pickerStyle(.segmented)
-                }
-            }
-        }
-    }
-
-    private var moreHubCard: some View {
-        Group {
-            if !MobileDestinations.moreDestinations(
-                context: shell.activeRoleContext,
-                platform: shell.platformFeatures
-            ).isEmpty {
-                LMSCard {
-                    NavigationLink(value: MoreHubRoute()) {
-                        HStack(spacing: 12) {
-                            Image(systemName: "square.grid.2x2.fill")
-                                .font(.footnote.weight(.semibold))
-                                .foregroundStyle(LexturesTheme.accent(for: colorScheme))
-                            Text(L.text("mobile.ia.more.title"))
-                                .font(.subheadline.weight(.semibold))
-                                .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
-                            Spacer(minLength: 0)
-                            Image(systemName: "chevron.right")
-                                .font(.caption.weight(.semibold))
-                                .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme).opacity(0.6))
-                        }
-                    }
-                    .buttonStyle(.plain)
-                }
-            }
-        }
-    }
-
-    private var offlineSyncCard: some View {
-        LMSCard {
-            Text(L.text("mobile.profile.pendingSync"))
-                .font(LexturesTheme.displayFont(17))
-                .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
-            Text(L.plural("mobile.pendingSync.waiting", count: offline.pendingCount))
-                .font(.caption)
-                .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
-            ForEach(offline.outboxItems.filter {
-                $0.status == .queued || $0.status == .failed || $0.status == .conflict
-            }) { item in
-                Divider()
-                VStack(alignment: .leading, spacing: 6) {
-                    Text(item.label)
-                        .font(.subheadline.weight(.medium))
-                        .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
-                    OutboxStatusChip(status: item.status)
-                    if item.status == .failed || item.status == .conflict {
-                        Button(L.text("mobile.profile.retry")) {
-                            Task { await offline.retryOutboxItem(id: item.id, accessToken: session.accessToken) }
-                        }
-                        .font(.caption.weight(.semibold))
-                    }
-                }
-            }
-        }
     }
 
     private var localeCard: some View {
@@ -319,15 +243,6 @@ struct ProfileView: View {
         navigatedMoreDestination = destination
     }
 
-    @ViewBuilder
-    private func moreDestinationView(_ destination: MoreDestination) -> some View {
-        if destination == .library && shell.platformFeatures.libraryBrowseEnabled {
-            LibraryBrowseView()
-        } else {
-            MoreDestinationPlaceholder(destination: destination)
-        }
-    }
-
     private var accountCard: some View {
         LMSCard {
             Text(L.text("mobile.profile.account"))
@@ -340,52 +255,6 @@ struct ProfileView: View {
                 value: shell.profile?.email ?? session.userEmail ?? L.text("mobile.emDash"),
                 systemImage: "envelope"
             )
-        }
-    }
-
-    private var notificationsCard: some View {
-        LMSCard {
-            NavigationLink(value: NotificationsRoute()) {
-                HStack(spacing: 12) {
-                    Image(systemName: "bell.fill")
-                        .font(.footnote.weight(.semibold))
-                        .foregroundStyle(LexturesTheme.accent(for: colorScheme))
-                        .frame(width: 32, height: 32)
-                        .background(LexturesTheme.brandTeal.opacity(colorScheme == .dark ? 0.18 : 0.14))
-                        .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-                    VStack(alignment: .leading, spacing: 2) {
-                        Text(L.text("mobile.profile.notifications"))
-                            .font(.subheadline.weight(.semibold))
-                            .foregroundStyle(LexturesTheme.textPrimary(for: colorScheme))
-                        Text(
-                            shell.unreadNotifications > 0
-                                ? L.format("mobile.profile.unread", shell.unreadNotifications)
-                                : L.text("mobile.dashboard.caughtUp")
-                        )
-                        .font(.caption)
-                        .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme))
-                    }
-                    Spacer(minLength: 0)
-                    if shell.unreadNotifications > 0 {
-                        Text("\(shell.unreadNotifications)")
-                            .font(.caption.weight(.bold))
-                            .foregroundStyle(.white)
-                            .padding(.horizontal, 8)
-                            .padding(.vertical, 3)
-                            .background(LexturesTheme.coral)
-                            .clipShape(Capsule())
-                    }
-                    Image(systemName: "chevron.right")
-                        .font(.caption.weight(.semibold))
-                        .foregroundStyle(LexturesTheme.textSecondary(for: colorScheme).opacity(0.6))
-                        .flipsForRightToLeftLayoutDirection(true)
-                }
-                .contentShape(Rectangle())
-            }
-            .buttonStyle(.plain)
-            .simultaneousGesture(TapGesture().onEnded {
-                Task { await PushManager.shared.requestPermissionIfNeeded() }
-            })
         }
     }
